+++ CHANGES.md
@@ -2,6 +2,10 @@ v0.0.4 (dev)
 ============
 + small bug-fixes
 + duphold now uses about half as much memory
++ add -d/--drop flag which will drop all samples from a VCF except the
+  one matching the sample in --bam. this simplifies per-sample 
+  parallelization followed by merge.
+
 
 v0.0.3
 ======